refactor: extract constants, separate API requests, zod validation
