+++ src/constans.tsx
@@ -1,0 +1,4 @@
+
+export const REDDIT_BASE_URL = "https://www.reddit.com"
+export const SUBREDDITS_PATH = "/subreddits.json";
+export const DEFAULT_IMAGE = "https://upload.wikimedia.org/wikipedia/commons/a/a3/Image-not-found.png?20210521171500";
