--- src/global.d.tsx
@@ -1,19 +1,0 @@
-declare global {
-
-	interface Subreddit {
-		title: string;
-		description: string;
-		image: string;
-		url: string;
-	  }
-	  
-	  interface Thread {
-		title: string;
-		author: string;
-		link: string;
-		text: string;
-	  }
-
-}
-
-export {}
