+++ src/types/subredditTypes.ts
@@ -1,0 +1,7 @@
+export interface Subreddit {
+	title: string;
+	description: string;
+	image: string;
+	url: string;
+}
+
