+++ src/types/threadTypes.ts
@@ -1,0 +1,7 @@
+export interface Thread {
+	title: string;
+	author: string;
+	link: string;
+	text: string;
+}
+
